requirements.txt
certifi
charset-normalizer
idna
requests
urllib3
confluent-kafka
influxdb_client
unicorn_binance_websocket_api
websockets==8.1